Refactor domain handling into mixin
