+++ pylintrc
@@ -321,13 +321,13 @@ max-statements=50
 max-parents=7
 
 # Maximum number of attributes for a class (see R0902).
-max-attributes=8
+max-attributes=7
 
 # Minimum number of public methods for a class (see R0903).
 min-public-methods=2
 
 # Maximum number of public methods for a class (see R0904).
-max-public-methods=30
+max-public-methods=22
 
 
 [CLASSES]
